grad
add user id

+++ GRAD_2022/lib/online/models/productdetailsmodel.dart
@@ -28,7 +28,7 @@ class Productdetailsmodel {
 
   int? id;
   String? name;
-  int? cost;
+  double? cost;
   int? discount;
   String? image;
   String? rfidCode;
